requirements.txt
anndata
easydl==2.2.0
faiss-cpu==1.7.4
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
POT==0.8.1.0
scanpy
scikit_learn==1.0.2
scipy==1.7.3
tensorboardX==2.6.2.2
tqdm==4.32.1
